re_search

Matches the first pattern in string.

Usage

re_search(pattern, string, ...)

Arguments

pattern

Pattern to look for.

The default interpretation is a regular expression, as described in vignette("regular-expressions"). Use regex() for finer control of the matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using fixed(). This is fast, but approximate. Generally, for matching human text, you'll want coll() which respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with boundary(). An empty pattern, "", is equivalent to boundary("character").

string

Input vector. Either a character vector, or something coercible to one.

...

Arguments passed on to re_compile

IGNORECASE

boolean Whether or not to ignore case during matching.

I

boolean Alternative to IGNORECASE.

MULTILINE

boolean whether or not the matches should span over multiple lines.

M

boolean ALternative for MULTILINE.

DOTALL

boolean Whether . should match line breaks.

S

boolean Alternative for DOTALL.

VERBOSE

boolean Whether pattern definitions span multiple lines.

X

boolean Alternative for VERBOSE.

NOFLAG

boolean If this flag is passed, all others are ignored.